project refactor part 2 done - extremums unified and more
