+++ console-app/src/main/java/com/isa/aem/MenuProject.java
@@ -1,16 +1,20 @@
 package com.isa.aem;
 
-import com.isa.aem.currency.calculator.ComplexConsoleCalculator;
-import com.isa.aem.global.extremum.GlobalExtremum;
-import com.isa.aem.local.extremum.LocalExtremumConsoleHandler;
-import com.isa.aem.tools.ConsolePrinter;
+import com.isa.aem.currency_calculator.ComplexConsoleCalculator;
+import com.isa.aem.data_loaders.FileContentReader;
+import com.isa.aem.helpers.ConsolePrinter;
+import com.isa.aem.rate_extremums.GlobalExtremumConsoleHandler;
+import com.isa.aem.rate_extremums.LocalExtremumConsoleHandler;
 
 import java.util.Scanner;
 
 public class MenuProject {
 
-    ConsolePrinter consolePrinter = new ConsolePrinter();
-    LocalExtremumConsoleHandler localExtremumConsoleHandler = new LocalExtremumConsoleHandler();
+    private ConsolePrinter consolePrinter = new ConsolePrinter();
+    private LocalExtremumConsoleHandler localExtremumConsoleHandler = new LocalExtremumConsoleHandler();
+    private GlobalExtremumConsoleHandler globalExtremumConsoleHandler = new GlobalExtremumConsoleHandler();
+    private MenuInformation menuInformation = new MenuInformation();
+    private ComplexConsoleCalculator calculator = new ComplexConsoleCalculator();
 
     public MenuProject() {
         FileContentReader fileContentReader = new FileContentReader();
@@ -18,17 +22,13 @@ public class MenuProject {
         fileContentReader.addPLNToListCurrency();
     }
 
-    MenuInformation menuInformation = new MenuInformation();
-    private GlobalExtremum globalExtremeMenu = new GlobalExtremum();
-    private ComplexConsoleCalculator calculator = new ComplexConsoleCalculator();
-
     public void menuPanel() {
         menuInformation.readMenu();
         pick();
     }
 
     private void pick() {
-        System.out.println("Wybierz interesującą Cię metodę");
+        System.out.println("Wybierz interesującą Cię opcję");
 
         do {
             Scanner scanner = new Scanner(System.in);
@@ -45,8 +45,7 @@ public class MenuProject {
                         calculator.run();
                         break;
                     case 3:
-                        System.out.println("\n\n\n\n     ");
-                        globalExtremeMenu.run();
+                        globalExtremumConsoleHandler.run();
                         break;
                     case 4:
                         consolePrinter.printLocalExtremumWelcome();
