+++ web-app/src/main/java/com/isa/aem/servlets/CurrencyManagerServlet.java
@@ -1,9 +1,12 @@
 package com.isa.aem.servlets;
 
 import com.isa.aem.*;
-import com.isa.aem.calculatorMethod.CreateAListOfAvailableCurrencies;
-import com.isa.aem.calculatorMethod.Score;
-import com.isa.aem.calculatorMethod.ScoreResult;
+import com.isa.aem.currency_calculator.CreateAListOfAvailableCurrencies;
+import com.isa.aem.currency_calculator.Score;
+import com.isa.aem.currency_calculator.ScoreResult;
+import com.isa.aem.data_loaders.CurrencyNameCountryFlagsLoader;
+import com.isa.aem.data_loaders.FileContentReader;
+import com.isa.aem.data_loaders.PropertiesLoader;
 import com.isa.aem.freemarker.TemplateName;
 import com.isa.aem.freemarker.TemplateProvider;
 import freemarker.template.Template;
@@ -45,14 +48,14 @@ public class CurrencyManagerServlet extends HttpServlet {
     @Inject
     private TemplateProvider templateProvider;
     public FileContentReader fileContentReader;
-    public LoadCurrencyNameCountryFlags loadCurrencyNameCountryFlags;
+    public CurrencyNameCountryFlagsLoader currencyNameCountryFlagsLoader;
 
     @Override
     public void init() throws ServletException {
         fileContentReader = new FileContentReader();
         fileContentReader.readFile();
         fileContentReader.addPLNToListCurrency();
-        loadCurrencyNameCountryFlags = new LoadCurrencyNameCountryFlags();
+        currencyNameCountryFlagsLoader = new CurrencyNameCountryFlagsLoader();
 
         AppProperties appProperties = PropertiesLoader.loadProperties();
         defaultCurrencyNameHave = appProperties.getCurrencyNamePln();
@@ -77,12 +80,12 @@ public class CurrencyManagerServlet extends HttpServlet {
         }
 
         if (score.getDateExchange() == null) {
-            LocalDate dateHaveMax = currencyRepository.getMostRecentDateForChosenCurrencyName(defaultCurrencyNameHave);
+            LocalDate dateHaveMax = currencyRepository.getNewestDateForChosenCurrencyName(defaultCurrencyNameHave);
             score.setDateExchange(dateHaveMax);
         }
 
         if (score.getMaxDate() == null) {
-            score.setMaxDate(currencyRepository.getMostRecentDateForChosenCurrencyName(defaultCurrencyNameHave));
+            score.setMaxDate(currencyRepository.getNewestDateForChosenCurrencyName(defaultCurrencyNameHave));
         }
 
         if (score.getMinDate() == null) {
@@ -128,7 +131,7 @@ public class CurrencyManagerServlet extends HttpServlet {
             String haveCurrency = calculatorCurrencyHaveTable[0];
             LocalDate date = score.scoreDate(reqDate, haveCurrency, calculatorCurrencyWantTable[0]);
             score = scoreResult.getScoreResult(haveCurrency, calculatorCurrencyWantTable[0], date, calculatorAmount);
-            score.setMaxDate(currencyRepository.getMostRecentDateForChosenCurrencyName(haveCurrency));
+            score.setMaxDate(currencyRepository.getNewestDateForChosenCurrencyName(haveCurrency));
             score.setMinDate(currencyRepository.getOldestDateForChosenCurrencyName(haveCurrency));
 
         } else if (ACTION_BUTTON_RANGE_CURRENCY.equals(action)) {
