+++ web-app/src/main/java/com/isa/aem/servlets/LocalExtremumServlet.java
@@ -1,10 +1,16 @@
 package com.isa.aem.servlets;
 
-import com.isa.aem.*;
-import com.isa.aem.calculatorMethod.Score;
+import com.isa.aem.AppProperties;
+import com.isa.aem.Currency;
+import com.isa.aem.CurrencyNameCountryFlags;
+import com.isa.aem.CurrencyRepository;
+import com.isa.aem.data_loaders.CurrencyNameCountryFlagsLoader;
+import com.isa.aem.data_loaders.FileContentReader;
+import com.isa.aem.data_loaders.PropertiesLoader;
 import com.isa.aem.freemarker.TemplateName;
 import com.isa.aem.freemarker.TemplateProvider;
-import com.isa.aem.local.extremum.LocalExtremum;
+import com.isa.aem.rate_extremums.ExtremumHelper;
+import com.isa.aem.rate_extremums.LocalExtremum;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
 
@@ -23,9 +29,9 @@ import java.util.Map;
 @WebServlet("/local-extremum")
 public class LocalExtremumServlet extends HttpServlet {
 
-    private Score score = new Score();
     private CurrencyRepository currencyRepository = new CurrencyRepository();
     private LocalExtremum localExtremum = new LocalExtremum();
+    private ExtremumHelper extremumHelper = new ExtremumHelper();
     CurrencyNameCountryFlags currencyNameCountryFlags = new CurrencyNameCountryFlags();
     private LocalDate dateFrom;
     private LocalDate dateTo;
@@ -42,13 +48,13 @@ public class LocalExtremumServlet extends HttpServlet {
     @Inject
     private TemplateProvider templateProvider;
     public FileContentReader fileContentReader;
-    public LoadCurrencyNameCountryFlags loadCurrencyNameCountryFlags;
+    public CurrencyNameCountryFlagsLoader currencyNameCountryFlagsLoader;
 
     @Override
     public void init() throws ServletException {
         fileContentReader = new FileContentReader();
         fileContentReader.readFile();
-        loadCurrencyNameCountryFlags = new LoadCurrencyNameCountryFlags();
+        currencyNameCountryFlagsLoader = new CurrencyNameCountryFlagsLoader();
 
         AppProperties appProperties = PropertiesLoader.loadProperties();
         defaultCurrencyName = appProperties.getCurrencyNameEur();
@@ -71,14 +77,14 @@ public class LocalExtremumServlet extends HttpServlet {
         }
 
         if (dateFrom == null) {
-            dateFrom = currencyRepository.getMostRecentDateMinusOneMonthForChosenCurrencyName(currencyName);
+            dateFrom = currencyRepository.getNewestDateMinusOneMonthForChosenCurrencyName(currencyName);
         }
 
         if (dateTo == null) {
-            dateTo = currencyRepository.getMostRecentDateForChosenCurrencyName(currencyName);
+            dateTo = currencyRepository.getNewestDateForChosenCurrencyName(currencyName);
         }
 
-        isDateFromAfterDateTo = localExtremum.isDateFromAfterDateTo(dateFrom, dateTo);
+        isDateFromAfterDateTo = extremumHelper.isDateFromAfterDateTo(dateFrom, dateTo);
         if (!isDateFromAfterDateTo) {
             minExtremum = localExtremum.getMinExtremum(currencyName, dateFrom, dateTo);
             maxExtremum = localExtremum.getMaxExtremum(currencyName, dateFrom, dateTo);
@@ -114,7 +120,7 @@ public class LocalExtremumServlet extends HttpServlet {
         dateFrom = LocalDate.parse(req.getParameter(DATE_FROM_PARAMETER));
         dateTo = LocalDate.parse(req.getParameter(DATE_TO_PARAMETER));
 
-        isDateFromAfterDateTo = localExtremum.isDateFromAfterDateTo(dateFrom, dateTo);
+        isDateFromAfterDateTo = extremumHelper.isDateFromAfterDateTo(dateFrom, dateTo);
         if (!isDateFromAfterDateTo) {
             minExtremum = localExtremum.getMinExtremum(currencyName, dateFrom, dateTo);
             maxExtremum = localExtremum.getMaxExtremum(currencyName, dateFrom, dateTo);
